requirements.txt
openai
requests
diffusers
transformers
torch
Pillow